making copy writable
Recent change that I've introduced to change the way input data is staged causes the input files to be readonly. If you copy it to local directory, the file attributes will also be set to read-only, so you won't be able to modify it. This change allows the App to modify the copied files.